+++ app_links/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 6.2.1
 * fix(Android): Add ProGuard rules.
+* fix(Windows): Send link if only there's 1 argument with a scheme.
 * chore(Android): Upgrade to latest AGP 8 and compileSdk 34.
 * chore: Update dependencies to latest.
 * chore(Android): Update example setup.